requirements.txt
flask
pandas
jinja2
gunicorn==21.2.0
python-dotenv
matplotlib
pathlib